Export env vars at ESP-IDF level
Things like `idf.py openocd` expects OPENOCD_SCRIPTS to be set in its
environment. Variables like these are defined in `tools.json` under the
`export_vars` key. They were previously being set in the environment of
the tool itself only, but it seems the ESP-IDF tooling expects them to
be set for everything.

So, instead of setting them in the tool binary wrappers, we communicate
them back up to the ESP-IDF derivation, which sets them in its setup
hook, for all downstream derivations to inherit.

See https://github.com/mirrexagon/nixpkgs-esp-dev/issues/61, https://github.com/mirrexagon/nixpkgs-esp-dev/issues/62

As a bonus, the wrapping was simplified back to about how it was before
macOS support, with a different workaround for the toolchain wrapper
issue (expects argv[0] to not have "-unwrapped" in it).

See https://github.com/mirrexagon/nixpkgs-esp-dev/pull/47#issuecomment-2023851486